Created Live plot class
